+++ edit-me/config/Config.d.ts
@@ -1,8 +1,0 @@
-import accents from '../../src/tokens/accents';
-import neutrals from '../../src/tokens/neutrals';
-
-interface ResumeConfig {
-  accentColor: keyof typeof accents;
-  neutralColor: keyof typeof neutrals;
-  ogImageTheme: 'dark' | 'light';
-}