+++ edit-me/config/resumeConfig.js
@@ -1,8 +1,0 @@
-/** @type {import('./Config').ResumeConfig} */
-const resumeConfig = {
-  accentColor: 'blue',
-  neutralColor: 'mauve',
-  ogImageTheme: 'light',
-};
-
-module.exports = resumeConfig;