+++ edit-me/config/resumeConfig.ts
@@ -1,0 +1,7 @@
+import { AccentColors, NeutralColors, ResumeConfig } from './Config';
+
+export default {
+  accentColor: AccentColors.Blue,
+  neutralColor: NeutralColors.Mauve,
+  ogImageTheme: 'light',
+} satisfies ResumeConfig;