+++ src/tokens/color.js
@@ -1,3 +1,0 @@
-module.exports = {
-  baseColors: { black: 'rgb(0, 0, 0)', white: 'rgb(255, 255, 255)' },
-};