+++ src/tokens/index.js
@@ -1,9 +1,0 @@
-const { baseColors } = require('./color');
-const accents = require('./accents');
-const neutrals = require('./neutrals');
-
-module.exports = {
-  accents,
-  baseColors,
-  neutrals,
-};